add ime support drop the guac keyboard
