--- addons/gst-web-core/lib/guacamole.js
@@ -1,2 +1,0 @@
-import Guacamole from './guacamole-keyboard-selkies.js';
-export default Guacamole;
